+++ elements/reactions/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/reactions
 
+## 21.2.5
+
+### Patch Changes
+
+- [`a424e62b264`](https://bitbucket.org/atlassian/atlassian-frontend/commits/a424e62b264) - Changes to support Node 16 Typescript definitions from `@types/node`.
+
 ## 21.2.4
 
 ### Patch Changes