+++ media/media-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-client
 
+## 15.1.1
+
+### Patch Changes
+
+- [`a424e62b264`](https://bitbucket.org/atlassian/atlassian-frontend/commits/a424e62b264) - Changes to support Node 16 Typescript definitions from `@types/node`.
+
 ## 15.1.0
 
 ### Minor Changes