+++ media/media-test-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-test-helpers
 
+## 29.4.2
+
+### Patch Changes
+
+- [`a424e62b264`](https://bitbucket.org/atlassian/atlassian-frontend/commits/a424e62b264) - Changes to support Node 16 Typescript definitions from `@types/node`.
+
 ## 29.4.1
 
 ### Patch Changes